+++ NEWS.md
@@ -1,3 +1,6 @@
+* Fixed error message when `require.package` is called from an anonymous
+  function and fails to load a package.  (Using `deparse(nlines = 1)` now.)
+
 v0.5-1 (2014-03-17)
 ===