+++ NEWS.md
@@ -1,6 +1,9 @@
 * 0.6-0
 * Added .add.extension. This allows users to create custom readers for extensions, either locally in a project or as packages.
 
+* Fixed error message when `require.package` is called from an anonymous
+  function and fails to load a package.  (Using `deparse(nlines = 1)` now.)
+
 v0.5-1 (2014-03-17)
 ===